Refactor MCP Frontend: Move to nvidia-nat-mcp package (#1332)
**Summary**
Refactored MCP frontend/server code from the core nvidia-nat package to the nvidia-nat-mcp package, following the same structure as the nvidia-nat-a2a package. 

**Dependencies**
This change removes the MCP SDK dependency from core and upgrades from MCP 1.13/1.14 to 1.25.

**New Directory structure**
```
packages/nvidia_nat_mcp/src/nat/plugins/mcp/
├── client/              # Client code (moved from top level)
│   ├── client_base.py
│   ├── client_config.py
│   └── client_impl.py
├── server/              # Frontend/server code (moved from src/nat/front_ends/mcp/)
│   ├── front_end_config.py
│   ├── front_end_plugin.py
│   ├── front_end_plugin_worker.py
│   ├── introspection_token_verifier.py
│   ├── memory_profiler.py
│   ├── register_frontend.py
│   └── tool_converter.py
├── auth/                # Auth code (unchanged)
├── tool.py              # MCP tool wrapper (unchanged)
└── utils.py             # Shared utilities (unchanged)

tests/nat/mcp/  
├── client/              # Client tests 
└── server/              # Server/frontend tests 
```

**Breaking changes**
1. This PR also drops the mcp_tool_wrapper function. I had put out a deprecation notice in the previous release recommending migration to mcp_client. Since then additional enhancements/fixes have been done to mcp_client which are not available in the depreceated mcp_tool_wrapper.
2. Folks using a custom MCP frontend worker will need to change the import path to `nat.plugins.mcp.server.front_end_plugin_worker.MCPFrontEndPluginWorker` from the previous `nat.front_ends.mcp.mcp_front_end_plugin_worker.MCPFrontEndPluginWorker`

## By Submitting this PR I confirm:
- I am familiar with the [Contributing Guidelines](https://github.com/NVIDIA/NeMo-Agent-Toolkit/blob/develop/docs/source/resources/contributing/index.md).
- We require that all contributors "sign-off" on their commits. This certifies that the contribution is your original work, or you have rights to submit it under the same license, or a compatible license.
  - Any contribution which contains commits that are not Signed-Off will not be accepted.
- When the PR is ready for review, new or existing tests cover these changes.
- When the PR is ready for review, the documentation is up to date with these changes.



## Summary by CodeRabbit

* **New Features**
  * Added MCP server registration entry point for plugin integration.

* **Chores**
  * Bumped MCP dependency to a newer compatible version and removed an older root dependency.
  * Added plugin entry point for server frontend.
  * Internal package/module reorganization and import relocations (non-functional).

* **Documentation**
  * Consolidated MCP docs to emphasize mcp_client usage; removed mcp_tool_wrapper examples and updated examples/notebooks.

* **Tests**
  * Updated tests to align with relocated modules and imports.

<sub>✏️ Tip: You can customize this high-level summary in your review settings.</sub>

Authors:
  - Anuradha Karuppiah (https://github.com/AnuradhaKaruppiah)

Approvers:
  - Will Killian (https://github.com/willkill07)
  - https://github.com/Salonijain27
  - David Gardner (https://github.com/dagardner-nv)

URL: https://github.com/NVIDIA/NeMo-Agent-Toolkit/pull/1332
