+++ examples/MCP/simple_calculator_mcp/README.md
@@ -50,9 +50,7 @@ You can publish the simple calculator tools using MCP using the `nat mcp serve` 
 ### MCP Client Configuration
 NeMo Agent toolkit enables workflows to use MCP tools as functions. The library handles the MCP server connection, tool discovery, and function registration. This allows the workflow to use MCP tools as regular functions.
 
-Tools served by remote MCP servers can be leveraged as NeMo Agent toolkit functions in one of two ways:
-- `mcp_client`: A flexible configuration using function groups, that allows you to connect to a MCP server, dynamically discover the tools it serves, and register them as NeMo Agent toolkit functions. `config-mcp-client.yml` example demonstrates how to use the `mcp_client` function group with both local and remote MCP servers.
-- `mcp_tool_wrapper`: A simple configuration that allows you to wrap a single MCP tool as a NeMo Agent toolkit function. `config-mcp-tool-wrapper.yml` example demonstrates how to use the `mcp_tool_wrapper` function group with a remote MCP server.
+Tools served by remote MCP servers can be leveraged as NeMo Agent toolkit functions using `mcp_client`, a flexible configuration using function groups that allows you to connect to an MCP server, dynamically discover the tools it serves, and register them as NeMo Agent toolkit functions. The `config-mcp-client.yml` example demonstrates how to use the `mcp_client` function group with both local and remote MCP servers.
 
 ### Running the example
 The `config-mcp-client.yml` example demonstrates how to use the `mcp_client` function group with both local and remote MCP servers. This configuration shows how to use multiple MCP servers with different transports in the same workflow.
