--- examples/MCP/simple_calculator_mcp/configs/config-mcp-tool-wrapper.yml
@@ -1,78 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2024-2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# This config file shows how to use the MCP server to access the `math` tools.
-# Here the workflow acts as a MCP client and connects to the MCP server running
-# on the specified URL. streamable-http is the recommended transport for HTTP-based
-# connections, but sse is also supported for backwards compatibility.
-
-
-functions:
-  calculator_add:
-    _type: mcp_tool_wrapper
-    url: "http://localhost:9901/mcp"
-    transport: "streamable-http"
-    mcp_tool_name: calculator__add
-    description: "Returns the sum of two numbers"
-  calculator_subtract:
-    _type: mcp_tool_wrapper
-    url: "http://localhost:9901/mcp"
-    transport: "streamable-http"
-    mcp_tool_name: calculator__subtract
-    description: "Returns the difference of two numbers"
-  calculator_multiply:
-    _type: mcp_tool_wrapper
-    url: "http://localhost:9901/mcp"
-    transport: "streamable-http"
-    mcp_tool_name: calculator__multiply
-    description: "Returns the product of two numbers"
-  calculator_divide:
-    _type: mcp_tool_wrapper
-    url: "http://localhost:9901/mcp"
-    transport: "streamable-http"
-    mcp_tool_name: calculator__divide
-    description: "Returns the quotient of two numbers"
-  calculator_compare:
-    _type: mcp_tool_wrapper
-    url: "http://localhost:9901/mcp"
-    transport: "streamable-http"
-    mcp_tool_name: calculator__compare
-    description: "Returns the comparison of two numbers"
-  current_datetime:
-    _type: current_datetime
-
-llms:
-  nim_llm:
-    _type: nim
-    model_name: meta/llama-3.1-70b-instruct
-    temperature: 0.0
-    max_tokens: 1024
-  openai_llm:
-    _type: openai
-    model_name: gpt-3.5-turbo
-    max_tokens: 2000
-
-workflow:
-  _type: react_agent
-  tool_names:
-    - calculator_add
-    - calculator_subtract
-    - calculator_multiply
-    - calculator_divide
-    - calculator_compare
-    - current_datetime
-  llm_name: nim_llm
-  verbose: true
-  parse_agent_response_max_retries: 3
