+++ packages/nvidia_nat_mcp/src/nat/plugins/mcp/client/client_impl.py
@@ -28,9 +28,9 @@ from nat.authentication.interfaces import AuthProviderBase
 from nat.builder.builder import Builder
 from nat.builder.function import FunctionGroup
 from nat.cli.register_workflow import register_function_group
-from nat.plugins.mcp.client_base import MCPBaseClient
-from nat.plugins.mcp.client_config import MCPClientConfig
-from nat.plugins.mcp.client_config import MCPToolOverrideConfig
+from nat.plugins.mcp.client.client_base import MCPBaseClient
+from nat.plugins.mcp.client.client_config import MCPClientConfig
+from nat.plugins.mcp.client.client_config import MCPToolOverrideConfig
 from nat.plugins.mcp.utils import truncate_session_id
 
 logger = logging.getLogger(__name__)
@@ -348,7 +348,7 @@ class MCPFunctionGroup(FunctionGroup):
 
     async def _create_session_client(self, session_id: str) -> tuple[MCPBaseClient, asyncio.Event, asyncio.Task]:
         """Create a new MCP client instance for the session."""
-        from nat.plugins.mcp.client_base import MCPStreamableHTTPClient
+        from nat.plugins.mcp.client.client_base import MCPStreamableHTTPClient
 
         config = self._client_config
         if not config:
@@ -484,9 +484,9 @@ async def mcp_client_function_group(config: MCPClientConfig, _builder: Builder):
     Returns:
         The function group
     """
-    from nat.plugins.mcp.client_base import MCPSSEClient
-    from nat.plugins.mcp.client_base import MCPStdioClient
-    from nat.plugins.mcp.client_base import MCPStreamableHTTPClient
+    from nat.plugins.mcp.client.client_base import MCPSSEClient
+    from nat.plugins.mcp.client.client_base import MCPStdioClient
+    from nat.plugins.mcp.client.client_base import MCPStreamableHTTPClient
 
     # Resolve auth provider if specified
     auth_provider = None
