+++ packages/nvidia_nat_mcp/src/nat/plugins/mcp/register.py
@@ -16,7 +16,8 @@
 # flake8: noqa
 # isort:skip_file
 
-# Import any providers which need to be automatically registered here
+# Register client components
+from .client import client_impl
 
-from . import client_impl
-from . import tool
+# Register server/frontend components
+from .server import register_frontend
