+++ packages/nvidia_nat_mcp/src/nat/plugins/mcp/server/__init__.py
@@ -12,16 +12,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from collections.abc import AsyncIterator
-
-from nat.cli.register_workflow import register_front_end
-from nat.data_models.config import Config
-from nat.front_ends.mcp.mcp_front_end_config import MCPFrontEndConfig
-
-
-@register_front_end(config_type=MCPFrontEndConfig)
-async def register_mcp_front_end(config: MCPFrontEndConfig, full_config: Config) -> AsyncIterator:
-    from nat.front_ends.mcp.mcp_front_end_plugin import MCPFrontEndPlugin
-
-    yield MCPFrontEndPlugin(full_config=full_config)
+"""MCP server/frontend components."""
