+++ packages/nvidia_nat_mcp/src/nat/plugins/mcp/server/front_end_plugin_worker.py
@@ -20,9 +20,10 @@ from collections.abc import Mapping
 from typing import TYPE_CHECKING
 from typing import Any
 
-from mcp.server.fastmcp import FastMCP
 from starlette.exceptions import HTTPException
 from starlette.requests import Request
+
+from mcp.server.fastmcp import FastMCP
 
 if TYPE_CHECKING:
     from fastapi import FastAPI
@@ -32,8 +33,8 @@ from nat.builder.function_base import FunctionBase
 from nat.builder.workflow import Workflow
 from nat.builder.workflow_builder import WorkflowBuilder
 from nat.data_models.config import Config
-from nat.front_ends.mcp.mcp_front_end_config import MCPFrontEndConfig
-from nat.front_ends.mcp.memory_profiler import MemoryProfiler
+from nat.plugins.mcp.server.front_end_config import MCPFrontEndConfig
+from nat.plugins.mcp.server.memory_profiler import MemoryProfiler
 from nat.runtime.session import SessionManager
 
 logger = logging.getLogger(__name__)
@@ -137,7 +138,7 @@ class MCPFrontEndPluginWorkerBase(ABC):
             mcp: The FastMCP server instance
             builder: The workflow builder instance
         """
-        from nat.front_ends.mcp.tool_converter import register_function_with_mcp
+        from nat.plugins.mcp.server.tool_converter import register_function_with_mcp
 
         # Set up the health endpoint
         self._setup_health_endpoint(mcp)
@@ -254,7 +255,7 @@ class MCPFrontEndPluginWorkerBase(ABC):
 
             from starlette.responses import JSONResponse
 
-            from nat.front_ends.mcp.tool_converter import get_function_description
+            from nat.plugins.mcp.server.tool_converter import get_function_description
 
             # Query params
             # Support repeated names and comma-separated lists
@@ -378,8 +379,9 @@ class MCPFrontEndPluginWorker(MCPFrontEndPluginWorkerBase):
         token_verifier = None
 
         if self.front_end_config.server_auth:
-            from mcp.server.auth.settings import AuthSettings
             from pydantic import AnyHttpUrl
+
+            from mcp.server.auth.settings import AuthSettings
 
             server_url = f"http://{self.front_end_config.host}:{self.front_end_config.port}"
             auth_settings = AuthSettings(issuer_url=AnyHttpUrl(self.front_end_config.server_auth.issuer_url),
@@ -387,7 +389,7 @@ class MCPFrontEndPluginWorker(MCPFrontEndPluginWorkerBase):
                                          resource_server_url=AnyHttpUrl(server_url))
 
             # Create token verifier
-            from nat.front_ends.mcp.introspection_token_verifier import IntrospectionTokenVerifier
+            from nat.plugins.mcp.server.introspection_token_verifier import IntrospectionTokenVerifier
 
             token_verifier = IntrospectionTokenVerifier(self.front_end_config.server_auth)
