+++ pyproject.toml
@@ -28,7 +28,6 @@ dependencies = [
   "httpx~=0.27",
   "jinja2~=3.1",
   "jsonpath-ng~=1.7",
-  "mcp~=1.13",
   "nest-asyncio2==1.7.1",
   "networkx~=3.4",
   # langchain-aws pins numpy to 1.* for Python <3.12 -- remove once package is updated
