+++ tests/nat/mcp/client/test_mcp_auth_provider.py
