+++ tests/nat/mcp/client/test_mcp_schema.py
