+++ tests/nat/mcp/server/test_mcp_debug_routes.py
@@ -21,8 +21,8 @@ from starlette.testclient import TestClient
 from nat.builder.function_base import FunctionBase
 from nat.data_models.config import Config
 from nat.data_models.config import GeneralConfig
-from nat.front_ends.mcp.mcp_front_end_config import MCPFrontEndConfig
-from nat.front_ends.mcp.mcp_front_end_plugin_worker import MCPFrontEndPluginWorker
+from nat.plugins.mcp.server.front_end_config import MCPFrontEndConfig
+from nat.plugins.mcp.server.front_end_plugin_worker import MCPFrontEndPluginWorker
 
 
 # Test fixtures and mock classes
